Support unconstrain version specifiers when adding packages
