+++ pdm/cli/completions/pdm.bash
@@ -1,7 +1,7 @@
 # BASH completion script for pdm
 # Generated by pycomplete 0.3.2
 
-_pdm_8820d2c61f1a916c_complete()
+_pdm_25182a7ef85b840e_complete()
 {
     local cur script coms opts com
     COMPREPLY=()
@@ -29,7 +29,7 @@ _pdm_8820d2c61f1a916c_complete()
         case "$com" in
 
             (add)
-            opts="--dev --editable --global --help --no-editable --no-self --no-sync --project --save-compatible --save-exact --save-wildcard --section --update-eager --update-reuse --verbose"
+            opts="--dev --editable --global --help --no-editable --no-self --no-sync --project --save-compatible --save-exact --save-wildcard --section --unconstrained --update-eager --update-reuse --verbose"
             ;;
 
             (build)
@@ -123,4 +123,4 @@ _pdm_8820d2c61f1a916c_complete()
     fi
 }
 
-complete -o default -F _pdm_8820d2c61f1a916c_complete pdm
+complete -o default -F _pdm_25182a7ef85b840e_complete pdm
