+++ pdm/cli/completions/pdm.fish
@@ -1,7 +1,7 @@
 # FISH completion script for pdm
 # Generated by pycomplete 0.3.2
 
-function __fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand
+function __fish_pdm_7426f3abf02b4bb8_complete_no_subcommand
     for i in (commandline -opc)
         if contains -- $i add build cache completion config export import info init install list lock remove run search show sync update use
             return 1
@@ -11,32 +11,32 @@ function __fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand
 end
 
 # global options
-complete -c pdm -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -l help -d 'show this help message and exit'
-complete -c pdm -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -l ignore-python -d 'Ignore the Python path saved in the pdm.toml config'
-complete -c pdm -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -l pep582 -d 'Print the command line to be eval\'d by the shell'
-complete -c pdm -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -l verbose -d '-v for detailed output and -vv for more detailed'
-complete -c pdm -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -l version -d 'show the version and exit'
+complete -c pdm -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -l help -d 'show this help message and exit'
+complete -c pdm -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -l ignore-python -d 'Ignore the Python path saved in the pdm.toml config'
+complete -c pdm -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -l pep582 -d 'Print the command line to be eval\'d by the shell'
+complete -c pdm -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -l verbose -d '-v for detailed output and -vv for more detailed'
+complete -c pdm -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -l version -d 'show the version and exit'
 
 # commands
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a add -d 'Add package(s) to pyproject.toml and install them'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a build -d 'Build artifacts for distribution'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a cache -d 'Control the caches of PDM'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a completion -d 'Generate completion scripts for the given shell'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a config -d 'Display the current configuration'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a export -d 'Export the locked packages set to other formats'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a import -d 'Import project metadata from other formats'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a info -d 'Show the project information'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a init -d 'Initialize a pyproject.toml for PDM'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a install -d 'Install dependencies from lock file'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a list -d 'List packages installed in the current working set'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a lock -d 'Resolve and lock dependencies'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a remove -d 'Remove packages from pyproject.toml'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a run -d 'Run commands or scripts with local packages loaded'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a search -d 'Search for PyPI packages'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a show -d 'Show the package information'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a sync -d 'Synchronize the current working set with lock file'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a update -d 'Update package(s) in pyproject.toml'
-complete -c pdm -f -n '__fish_pdm_0e4faf7eb1ce3130_complete_no_subcommand' -a use -d 'Use the given python version or path as base interpreter'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a add -d 'Add package(s) to pyproject.toml and install them'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a build -d 'Build artifacts for distribution'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a cache -d 'Control the caches of PDM'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a completion -d 'Generate completion scripts for the given shell'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a config -d 'Display the current configuration'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a export -d 'Export the locked packages set to other formats'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a import -d 'Import project metadata from other formats'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a info -d 'Show the project information'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a init -d 'Initialize a pyproject.toml for PDM'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a install -d 'Install dependencies from lock file'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a list -d 'List packages installed in the current working set'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a lock -d 'Resolve and lock dependencies'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a remove -d 'Remove packages from pyproject.toml'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a run -d 'Run commands or scripts with local packages loaded'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a search -d 'Search for PyPI packages'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a show -d 'Show the package information'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a sync -d 'Synchronize the current working set with lock file'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a update -d 'Update package(s) in pyproject.toml'
+complete -c pdm -f -n '__fish_pdm_7426f3abf02b4bb8_complete_no_subcommand' -a use -d 'Use the given python version or path as base interpreter'
 
 # command options
 
@@ -53,6 +53,7 @@ complete -c pdm -A -n '__fish_seen_subcommand_from add' -l save-compatible -d 'S
 complete -c pdm -A -n '__fish_seen_subcommand_from add' -l save-exact -d 'Save exact version specifiers'
 complete -c pdm -A -n '__fish_seen_subcommand_from add' -l save-wildcard -d 'Save wildcard version specifiers'
 complete -c pdm -A -n '__fish_seen_subcommand_from add' -l section -d 'Specify target section to add into'
+complete -c pdm -A -n '__fish_seen_subcommand_from add' -l unconstrained -d 'Ignore the version constraint of pinned packages'
 complete -c pdm -A -n '__fish_seen_subcommand_from add' -l update-eager -d 'Try to update the packages and their dependencies recursively'
 complete -c pdm -A -n '__fish_seen_subcommand_from add' -l update-reuse -d 'Reuse pinned versions already present in lock file if possible'
 complete -c pdm -A -n '__fish_seen_subcommand_from add' -l verbose -d '-v for detailed output and -vv for more detailed'
