+++ pdm/resolver/providers.py
@@ -136,8 +136,10 @@ class ReusePinProvider(BaseProvider):
         if identifier not in self.tracked_names and identifier in self.preferred_pins:
             pin = self.preferred_pins[identifier]
             incompat = list(incompatibilities[identifier])
-            if pin not in incompat:
-                pin._preferred = True
+            pin._preferred = True
+            if pin not in incompat and all(
+                self.is_satisfied_by(r, pin) for r in requirements[identifier]
+            ):
                 yield pin
         yield from super().find_matches(identifier, requirements, incompatibilities)
